Made snitch move
I combined a vector pointing to the center of the play area with a random vector to create a somewhat smooth snitch that doesn't just fly to the edge of the arena and stick to a wall!

Drew a line with magnitude that shows the direction (and magnitude) of the snitch.